# src/openapi/api.schema.yml
openapi: 3.0.3
info:
  description: API for Slice Dice Service.
  version: 1.0.0
  title: Slice Dice Base API
...
paths:
...
  /auth/logout:
    post:
      tags:
        - SliceDice
      summary: Logout the user
      operationId: postLogout
      requestBody:
        description: 'logout'
        required: true
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/UserToken'
      responses:
        '201':
          description: 'User has been successfully logged out'
          content:

            application/json:
              schema:
                type: object
                required: [ 'id' ]
                properties:
                  id:
                    type: string
                    format: 'uuid'
                  revision:
                    type: string
          headers:
            Location:
              schema:
                type: string
                format: URI
        '400':
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        500:
          $ref: '#/components/responses/InternalServerError'
  /auth/update:
    put:
      tags:
        - SliceDice
      summary: update user
      operationId: updateUser
      requestBody:
        description: 'update the user profile'
        required: true
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/UserEntity'
      responses:
        '201':
          description: 'User has been successfully created'
          content:
            application/json:
              schema:
                type: object
                required: [ 'id' ]
                properties:
                  id:
                    type: string
                    format: 'uuid'
                  revision:
                    type: string
          headers:
            Location:
              schema:
                type: string
                format: URI
        '400':
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        500:
          $ref: '#/components/responses/InternalServerError'
...
  /auth/users:
    get:
      tags:
        - SliceDice
      summary: Get The Users
      operationId: getUsers
      responses:
        200:
          description: A list of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserList'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Unauthorized'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        500:
          $ref: '#/components/responses/InternalServerError'
...
  /auth/vehicle:
    get:
      tags:
        - SliceDice
      summary: Get The Vehicles
      operationId: getVehicles
      responses:
        200:
          description: A list of vehicles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleList'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Unauthorized'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        500:
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - SliceDice
      summary: Add The Vehicle
      operationId: postSaveVehicle
      requestBody:
        description: 'add vehicle'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Vehicle'
      responses:
        '201':
          description: 'Password reset link has been sent'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response'
          headers:
            Location:
              schema:
                type: string
                format: URI
        '400':
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        500:
          $ref: '#/components/responses/InternalServerError'
...
  /auth/setChargingProfile:
    post:
      tags:
        - ChargePoint
      summary: 5.16. Set Charging Profile
      operationId: SetChargingProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetChargingProfile'
      responses:
        '200':
          description: >
            This contains the field definition of the SetChargingProfile.conf PDU sent by the Charge Point to the Central System in response to a SetChargingProfile.req PDU.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SetChargingProfileResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '5XX':
          $ref: '#/components/responses/InternalError'
...
  /auth/setConnector:
    post:
      tags:
        - ChargePoint
      summary: 5.17. Set Connector Object
      operationId: SetConnector
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RateObject'
      responses:
        '200':
          description: >
            This contains the Connector Object details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '5XX':
          $ref: '#/components/responses/InternalError'
  /auth/chargeStation:
    post:
      tags:
        - ChargePoint
      summary: 5.18. Set Charge Station Object
      operationId: SetChargeStation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChargeStation'
      responses:
        '200':
          description: >
            This contains the ChargeStation Object details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '5XX':
          $ref: '#/components/responses/InternalError'

    get:
      tags:
        - ChargePoint
      summary: Get a list of connected charging stations.
      operationId: ChargePointList
      parameters:
        - in: query
          name: location
          description: 'location'
          schema:
            type: string
        - in: query
          name: model
          description: 'model'
          schema:
            type: string
        - in: query
          name: active
          description: 'active'
          schema:
            type: boolean
      responses:
        '200':
          description: Successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChargePointListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '5XX':
          $ref: '#/components/responses/InternalError'
components:
...
  responses:
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            code: 'error.resource.not-found'
            message: 'Requested resource with ID: 3fa85f64-5717-4562-b3fc-2c963f66afa6 was not found'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            code: 'error.auth.unauthorized'
            status: 401
    InternalError:
      description: Internal error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error'
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            status: 500
            message: Unexpected error while validating the request id
...
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            statusCode: 'error.request.invalid'
            message: 'Request contains invalid values or missing required properties'
...
    PreconditionFailed:
      description: Precondition Failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            code: 'error.auth.github-access-token-expired'
            status: 412
  schemas:
    VehicleList:
      type: array
      items:
        $ref: '#/components/schemas/Vehicle'
    Vehicle:
      type: object
      properties:
        brand:
          type: string
        model:
          type: string
        year:
          type: string
        pictureLink:
          type: string
        postCode:
          type: string
...
    UserList:
      type: array
      items:
        $ref: '#/components/schemas/UserEntity'
    UserEntity:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
        displayName:
          type: string
          nullable: true
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        avatar:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        password:
          type: string
          nullable: true
        salary:
          type: string
          nullable: true
        currency:
          type: string
          nullable: true
        department:
          type: string
          nullable: true
        on_contract:
          type: string
          nullable: true
        sub_department:
          type: string
          nullable: true
        addresses:
          type: array
          items:
            $ref: '#/components/schemas/AddressItem'
        privileges:
          type: array
          items:
            $ref: '#/components/schemas/PrivilegeItem'
    AddressItem:
      type: object
      properties:
        line1:
          type: string
        city:
          type: string
    PrivilegeItem:
      type: object
      properties:
        entity:
          type: string
        read:
          type: boolean
        create:
          type: boolean
        update:
          type: boolean
        delete:
          type: boolean
        admin:
          type: boolean
    UserToken:
      type: object
      additionalProperties: false
      properties:
        refreshToken:
          type: string
...
    ChargingSchedule:
      type: object
      required:
        - chargingRateUnit
      properties:
        duration:
          type: integer
          format: int32
          description: >
            Duration of the charging schedule in seconds.
            If the duration is left empty, the last period will continue indefinitely or until end of the transaction in case startSchedule is absent.
        startSchedule:
          type: string
          format: date-time
        chargingRateUnit:
          type: string
          enum: [ W, A ]
          description: The unit of measure Limit is expressed in.
        startPeriod:
          type: integer
          format: int32
          description: Start of the period, in seconds from the start of schedule. The value of StartPeriod also defines the stop time of the previous period.
        limit:
          type: number
          format: double
          description: Charging rate limit during the schedule period, in the applicable chargingRateUnit, for example in Amperes or Watts.
        numberPhases:
          type: integer
          format: int32
          description: The number of phases that can be used for charging. If a number of phases is needed, numberPhases=3 will be assumed unless another number is given.
        minChargingRate:
          type: number
          format: double
          description: >
            Minimum charging rate supported by the electric vehicle. The unit of measure is defined by the chargingRateUnit.
            This parameter is intended to be used by a local smart charging algorithm to optimize the power allocation for in the case a
            charging process is inefficient at lower charging rates.
    ChargingProfile:
      type: object
      required:
        - chargingProfileId
        - stackLevel
        - chargingProfilePurpose
        - chargingProfileKind
        - chargingSchedule
      properties:
        chargingProfileId:
          type: integer
          format: int32
          description: Unique identifier for this profile.
        transactionId:
          type: integer
          format: int32
          description: Only valid if ChargingProfilePurpose is set to TxProfile, the transactionId MAY be used to match the profile to a specific transaction.
        stackLevel:
          type: integer
          format: int32
          description: Value determining level in hierarchy stack of profiles. Higher values have precedence over lower values. Lowest level is 0.
        chargingProfilePurpose:
          type: string
          description: Defines the purpose of the schedule transferred by this message.
          enum: [ ChargePointMaxProfile, TxDefaultProfile, TxProfile ]
        chargingProfileKind:
          type: string
          enum: [ Absolute, Recurring, Relative ]
          description: Indicates the kind of schedule.
        recurrencyKind:
          type: string
          enum: [ Daily, Weekly ]
          description: Indicates the start point of a recurrence.
        validFrom:
          type: string
          format: date-time
          description: Point in time at which the profile starts to be valid. If absent, the profile is valid as soon as it is received by the Charge Point.
        validTo:
          type: string
          format: date-time
          description: Point in time at which the profile stops to be valid. If absent, the profile is valid until it is replaced by another profile.
        chargingSchedule:
          $ref: '#/components/schemas/ChargingSchedule'
    SetChargingProfile:
      type: object
      required:
        - connectorId
        - csChargingProfiles
      properties:
        connectorId:
          type: integer
          format: int32
          description: >
            The connector to which the charging profile applies. If connectorId = 0,the message contains an overall limit for the Charge Point.
        csChargingProfiles:
          $ref: '#/components/schemas/ChargingProfile'
    SetChargingProfileResponse:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          description: >
            Returns whether the Charge Point has been able to process the message successfully.
            This does not guarantee the schedule will be followed to the letter. There might be other constraints the Charge Point may need to take into account.
          enum: [ Accepted, Rejected, NotSupported ]
    ChargePointListResponse:
      type: array
      items:
        $ref: '#/components/schemas/ChargeStation'
    PriceComponent:
      type: array
      items:
        type: object
        properties:
          tax:
            type: number
          step_size:
            type: number
          type:
            type: string
          price:
            type: number
    RateObject:
      type: object
      properties:
        currency:
          type: string
        price_components:
          $ref: '#/components/schemas/PriceComponent'
    Connector:
      type: object
      properties:
        active:
          type: boolean
        status:
          type: string
        type:
          type: string
        format:
          type: string
        power_type:
          type: string
        power:
          type: number
        chargestation:
          type: string
        rate:
          $ref: '#/components/schemas/RateObject'
    BootInfo:
      type: object
      properties:
        chargeBoxSerialNumber:
          type: string
        chargePointModel:
          type: string
        chargePointSerialNumber:
          type: string
        chargePointVendor:
          type: string
        firmwareVersion:
          type: string
        iccid:
          type: string
        imsi:
          type: string
    ChargeStation:
      type: object
      properties:
        location:
          type: string
        protocol:
          type: string
        endpoint:
          type: string
        static_endpoint:
          type: string
        online:
          type: boolean
        active:
          type: boolean
        public:
          type: boolean
        model:
          type: string
        bootInfo:
          $ref: '#/components/schemas/BootInfo'
        coordinates:
          type: array
          example: [4.865672799999999,
                    52.3310936]
        connectors:
          type: array
          items:
            $ref: '#/components/schemas/Connector'
        lastConnectAt:
          type: string
        lastDisconnectAt:
          type: string
        lastMessageAt:
          type: string
...
    ApiError:
      type: object
      properties:
        message:
          type: string
          description: A short, summary of the problem type. Written in english and readable for engineers.
          example: Service Unavailable
        code:
          type: string
          description: Error code. Can be used for translation.
          example: error.resource.not-found

        target:
          type: string
          description: The target of the error.

        details:
          type: array
          items:
            $ref: '#/components/schemas/ApiError'

          description: List of failed swagger validations.
          example: |
            [
              {
                "target": ".body.name",
                "message": "should match pattern \"^([a-zA-Z])([a-zA-Z]|\\d|\\ )*$\"",
                "code": "pattern.openapi.validation"
              },
              {
                "target": ".body.logo",
                "message": "should match format \"uuid\"",
                "code": "format.openapi.validation"
              }
            ]
    response:
      type: object
      required:
        - code
        - message
      properties:
        error:
          type: object
          properties:
            code:
              type: integer
              format: int32
            message:
              type: string
    error:
      type: object
      required:
        - code
        - message
      properties:
        error:
          type: object
          properties:
            code:
              type: integer
              format: int32
            message:
              type: string
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...